Splitted manage to add & remove files
