+++ portion/commands/add.py
@@ -6,7 +6,7 @@ from portion.core import TemplateManager
 from portion.models import Template
 
 
-class ManageCommand(CommandBase):
+class AddCommand(CommandBase):
     def __init__(self) -> None:
         super().__init__()
         self.project_manager = ProjectManager()
@@ -32,18 +32,3 @@ class ManageCommand(CommandBase):
 
         self.project_manager.update_configuration(path, config)
         self.logger.info(f"{template_name} has been added successfully")
-
-    def remove(self, template_name: str) -> None:
-        path = sys.path[0]
-        config = self.project_manager.read_configuration(path)
-
-        for i, template in enumerate(config.templates):
-            if template.name == template_name:
-                config.templates.pop(i)
-                break
-        else:
-            self.logger.error("The template isn't exist in this project")
-            return None
-
-        self.project_manager.update_configuration(path, config)
-        self.logger.info(f"{template_name} has been removed successfully")
