+++ portion/handler/__init__.py
@@ -1,8 +1,9 @@
 from typer import Typer
 
+from .add import AddHandler
 from .init import InitHandler
-from .manage import ManageHandler
 from .new import NewHandler
+from .remove import RemoveHandler
 from .template import TemplateHandler
 from .version import VersionHandler
 
@@ -13,7 +14,8 @@ def load_handlers(app: Typer) -> None:
         TemplateHandler,
         NewHandler,
         VersionHandler,
-        ManageHandler,
+        AddHandler,
+        RemoveHandler,
     ]
 
     for command in all_commands:
